Added a safety check and fixed a couple of warnings.

+++ modules/juce_audio_basics/midi/juce_MidiMessage.cpp
@@ -679,8 +679,8 @@ MidiMessage MidiMessage::textMetaEvent (int type, const StringRef& text)
 
     const size_t headerLen = sizeof (header) - n;
 
-    uint8* const dest = result.allocateSpace (headerLen + textSize);
-    result.size = headerLen + textSize;
+    uint8* const dest = result.allocateSpace ((int) (headerLen + textSize));
+    result.size = (int) (headerLen + textSize);
 
     memcpy (dest, header + n, headerLen);
     memcpy (dest + headerLen, text.text.getAddress(), textSize);
